Move hub tests into its own package and implement unpair
